test : DogControllerTest 의 로직을 VoteControllerTest 로 이동
