+++ src/test/java/com/dogpopularityvote/presentation/VoteControllerTest.java
@@ -1,0 +1,49 @@
+package com.dogpopularityvote.presentation;
+
+import com.dogpopularityvote.application.DogService;
+import com.dogpopularityvote.application.VoteProducerService;
+import com.dogpopularityvote.dto.request.DogVoteRequest;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.willDoNothing;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@WebMvcTest(VoteController.class)
+class VoteControllerTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Autowired
+    private ObjectMapper objectMapper;
+
+    @MockBean
+    private VoteProducerService voteProducerService;
+
+    @Test
+    void 강아지에_투표를_한다() throws Exception {
+        // given
+        DogVoteRequest request = new DogVoteRequest();
+
+        willDoNothing().given(voteProducerService).sendVote(any(DogVoteRequest.class));
+
+        // when & then
+        mockMvc.perform(post("/api/votes")
+                        .accept(MediaType.APPLICATION_JSON)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(request))
+                )
+                .andDo(print())
+                .andExpect(status().isNoContent());
+    }
+}
